pyproject.toml
[project]
name = "gatp"
version = "0.1.0"
description = "Adaptive Topological Policies for Git Repositories"
authors = [
    { name = "Matteo Breschi" }
]
dependencies = [
    "typer",
    "GitPython",
    "requests"
]

[project.scripts]
gatp = "gatp.app:app"
